refactor : ValidationUtil 클래스 삭제에 따른 유효성 체크 로직 이동
